+++ test/EcPayClientTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use CarlLee\EcPayB2C\Contracts\CommandInterface;
+use CarlLee\EcPayB2C\Contracts\EncryptableCommandInterface;
 use CarlLee\EcPayB2C\EcPayClient;
 use CarlLee\EcPayB2C\Infrastructure\CipherService;
 use CarlLee\EcPayB2C\Infrastructure\PayloadEncoder;
@@ -71,16 +71,16 @@ class EcPayClientTest extends TestCase
         $ecPayClient = new EcPayClient($this->server, $wrongKey, $wrongIV);
 
         // Mock Command
-        $invoice = Mockery::mock(CommandInterface::class);
+        $invoice = Mockery::mock(EncryptableCommandInterface::class);
         $invoice->shouldReceive('setHashKey')->with($wrongKey);
         $invoice->shouldReceive('setHashIV')->with($wrongIV);
-        $invoice->shouldReceive('getPayload')->andReturn([
-            'MerchantID' => 'TEST_MERCHANT_ID',
-            'RqHeader' => ['Timestamp' => time()],
-            'Data' => ['Example' => 'value'],
-        ]);
         $invoice->shouldReceive('getPayloadEncoder')->andReturn(new PayloadEncoder(new CipherService($this->hashKey, $this->hashIV)));
         $invoice->shouldReceive('getRequestPath')->andReturn('/test/api');
+        $invoice->shouldReceive('getContent')->andReturn([
+            'MerchantID' => 'TEST_MERCHANT_ID',
+            'RqHeader' => ['Timestamp' => time()],
+            'Data' => $this->encryptData(['Example' => 'value']),
+        ]);
 
         $response = $ecPayClient->send($invoice);
 
@@ -109,16 +109,16 @@ class EcPayClientTest extends TestCase
 
         $ecPayClient = new EcPayClient($this->server, $this->hashKey, $this->hashIV);
 
-        $invoice = Mockery::mock(CommandInterface::class);
+        $invoice = Mockery::mock(EncryptableCommandInterface::class);
         $invoice->shouldReceive('setHashKey');
         $invoice->shouldReceive('setHashIV');
-        $invoice->shouldReceive('getPayload')->andReturn([
-            'MerchantID' => 'TEST_MERCHANT_ID',
-            'RqHeader' => ['Timestamp' => time()],
-            'Data' => [],
-        ]);
         $invoice->shouldReceive('getPayloadEncoder')->andReturn(new PayloadEncoder(new CipherService($this->hashKey, $this->hashIV)));
         $invoice->shouldReceive('getRequestPath')->andReturn('/test/api');
+        $invoice->shouldReceive('getContent')->andReturn([
+            'MerchantID' => 'TEST_MERCHANT_ID',
+            'RqHeader' => ['Timestamp' => time()],
+            'Data' => $this->encryptData([]),
+        ]);
 
         $response = $ecPayClient->send($invoice);
         $data = $response->getData();
@@ -146,16 +146,16 @@ class EcPayClientTest extends TestCase
 
         $ecPayClient = new EcPayClient($this->server, $this->hashKey, $this->hashIV);
 
-        $invoice = Mockery::mock(CommandInterface::class);
+        $invoice = Mockery::mock(EncryptableCommandInterface::class);
         $invoice->shouldReceive('setHashKey');
         $invoice->shouldReceive('setHashIV');
-        $invoice->shouldReceive('getPayload')->andReturn([
-            'MerchantID' => 'TEST_MERCHANT_ID',
-            'RqHeader' => ['Timestamp' => time()],
-            'Data' => [],
-        ]);
         $invoice->shouldReceive('getPayloadEncoder')->andReturn(new PayloadEncoder(new CipherService($this->hashKey, $this->hashIV)));
         $invoice->shouldReceive('getRequestPath')->andReturn('/test/api');
+        $invoice->shouldReceive('getContent')->andReturn([
+            'MerchantID' => 'TEST_MERCHANT_ID',
+            'RqHeader' => ['Timestamp' => time()],
+            'Data' => $this->encryptData([]),
+        ]);
 
         $this->expectException(Exception::class);
         // Note: EcPayClient throws generic Exception for json error after decryption fail (decryption returns false/garbage)